feat(core): move register to plugins/js
