+++ packages/devkit/src/generators/to-js.ts
@@ -1,6 +1,6 @@
 import type { Tree } from 'nx/src/generators/tree';
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { typescriptVersion } from 'nx/src/utils/versions';
+import { typescriptVersion } from '../utils/versions';
 import { ensurePackage } from '../utils/package-json';
 
 /**
