+++ packages/devkit/src/utils/versions.ts
@@ -1,0 +1,1 @@
+export const typescriptVersion = '~4.9.5';
