+++ packages/eslint-plugin-nx/src/resolve-workspace-rules.ts
@@ -1,6 +1,6 @@
 import type { TSESLint } from '@typescript-eslint/utils';
 import { existsSync } from 'fs';
-import { registerTsProject } from 'nx/src/utils/register';
+import { registerTsProject } from 'nx/src/plugins/js/utils/register';
 import { WORKSPACE_PLUGIN_DIR, WORKSPACE_RULE_NAMESPACE } from './constants';
 
 type ESLintRules = Record<string, TSESLint.RuleModule<string, unknown[]>>;
