+++ packages/js/src/utils/versions.ts
@@ -1,5 +1,3 @@
-export { swcCoreVersion } from 'nx/src/utils/versions';
-
 export const nxVersion = require('../../package.json').version;
 
 export const esbuildVersion = '^0.17.5';
@@ -8,3 +6,5 @@ export const swcCliVersion = '~0.1.55';
 export const swcHelpersVersion = '~0.4.11';
 export const typesNodeVersion = '18.7.1';
 export const typescriptVersion = '~4.9.5';
+
+export const swcCoreVersion = '^1.2.173';
