+++ packages/linter/src/generators/workspace-rules-project/workspace-rules-project.ts
@@ -16,7 +16,7 @@ import {
 import { getRelativePathToRootTsConfig } from '@nx/js';
 import { join } from 'path';
 import { workspaceLintPluginDir } from '../../utils/workspace-lint-rules';
-import { swcCoreVersion, swcNodeVersion } from 'nx/src/utils/versions';
+import { swcCoreVersion, swcNodeVersion } from '../../utils/versions';
 import { nxVersion } from '../../utils/versions';
 
 export const WORKSPACE_RULES_PROJECT_NAME = 'eslint-rules';
