+++ packages/linter/src/utils/versions.ts
@@ -1,7 +1,8 @@
 export const nxVersion = require('../../package.json').version;
 
-export const tslintToEslintConfigVersion = '^2.13.0';
-
-export const typescriptESLintVersion = '^5.58.0';
 export const eslintVersion = '~8.15.0';
 export const eslintConfigPrettierVersion = '8.1.0';
+export const swcNodeVersion = '^1.4.2';
+export const swcCoreVersion = '^1.2.173';
+export const tslintToEslintConfigVersion = '^2.13.0';
+export const typescriptESLintVersion = '^5.58.0';
