+++ packages/nx-plugin/src/generators/plugin/plugin.ts
@@ -12,7 +12,7 @@ import {
 import { libraryGenerator as jsLibraryGenerator } from '@nx/js';
 import { addSwcDependencies } from '@nx/js/src/utils/swc/add-swc-dependencies';
 import { Linter } from '@nx/linter';
-import { swcNodeVersion } from 'nx/src/utils/versions';
+import { swcNodeVersion } from '../../utils/versions';
 import * as path from 'path';
 
 import { nxVersion, tsLibVersion } from '../../utils/versions';
