+++ packages/nx-plugin/src/utils/versions.ts
@@ -1,3 +1,5 @@
 export const nxVersion = require('../../package.json').version;
 export const jsoncEslintParserVersion = '^2.1.0';
 export const tsLibVersion = '^2.3.0';
+export const swcNodeVersion = '^1.4.2';
+export const swcCoreVersion = '^1.2.173';
