+++ packages/nx/src/plugins/js/utils/register.ts
@@ -1,6 +1,6 @@
 import { dirname, join } from 'path';
 import type { CompilerOptions } from 'typescript';
-import { logger, NX_PREFIX, stripIndent } from './logger';
+import { logger, NX_PREFIX, stripIndent } from '../../../utils/logger';
 
 const swcNodeInstalled = packageIsInstalled('@swc-node/register');
 const tsNodeInstalled = packageIsInstalled('ts-node/register');
