+++ packages/nx/src/utils/register.spec.ts
@@ -1,5 +1,5 @@
 import { JsxEmit, ModuleKind, ScriptTarget } from 'typescript';
-import { getTsNodeCompilerOptions } from './register';
+import { getTsNodeCompilerOptions } from '../plugins/js/utils/register';
 
 describe('getTsNodeCompilerOptions', () => {
   it('should replace enum value with enum key for module', () => {
