+++ packages/nx/src/utils/versions.ts
@@ -1,10 +1,1 @@
 export const nxVersion = require('../../package.json').version;
-
-export const angularCliVersion = '~15.2.0';
-export const typescriptVersion = '~4.9.5';
-export const prettierVersion = '^2.6.2';
-export const typescriptESLintVersion = '^5.58.0';
-export const eslintVersion = '~8.15.0';
-export const eslintConfigPrettierVersion = '8.1.0';
-export const swcNodeVersion = '^1.4.2';
-export const swcCoreVersion = '^1.2.173';
