+++ packages/vite/src/executors/dev-server/dev-server.impl.ts
@@ -11,7 +11,7 @@ import {
 
 import { ViteDevServerExecutorOptions } from './schema';
 import { ViteBuildExecutorOptions } from '../build/schema';
-import { registerTsConfigPaths } from 'nx/src/utils/register';
+import { registerTsConfigPaths } from 'nx/src/plugins/js/utils/register';
 import { resolve } from 'path';
 
 export async function* viteDevServerExecutor(
